Version 2.2.5 release - added 3-byte parameters to SM command for really long moves.
git-svn-id: https://eggbotcode.googlecode.com/svn/trunk@303 72233254-1b6c-9e9c-5072-401df62706fb

+++ EBB_firmware/app.X/source/ebb.h
@@ -78,7 +78,7 @@ typedef struct
 {
     CommandType     Command;
     INT16           StepAdd[NUMBER_OF_STEPPERS];
-    UINT16          StepsCounter[NUMBER_OF_STEPPERS];
+    UINT32          StepsCounter[NUMBER_OF_STEPPERS];
     UINT8           DirBits;
     UINT32          DelayCounter;   // NOT Milliseconds! In 25KHz units
     UINT16          ServoPosition;
